refactor: update deps and simplify codes
