+++ app/[lang]/(hyperjump)/case-studies/[slug]/page.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { notFound } from "next/navigation";
 
 import { Button } from "@/components/ui/button";
-import { caseStudyButton, caseStudyMore } from "@/locales/.generated/server";
+import { caseStudyButton, caseStudyMore } from "@/locales/.generated/strings";
 import {
   supportedLanguages,
   type SupportedLanguage
