+++ app/[lang]/(hyperjump)/components/location.tsx
@@ -8,7 +8,7 @@ import { Button } from "@/components/ui/button";
 import {
   mainOpenInGoogleMaps,
   mainOurLocation
-} from "@/locales/.generated/server";
+} from "@/locales/.generated/strings";
 import type { SupportedLanguage } from "@/locales/.generated/types";
 
 type Location = {
