+++ app/[lang]/(inferenceai)/inferenceai/components/footer.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 
 import { Separator } from "@/components/ui/separator";
 import type { SupportedLanguage } from "@/locales/.generated/types";
-import { copyright, inferenceaiFooter } from "@/locales/.generated/server";
+import { copyright, inferenceaiFooter } from "@/locales/.generated/strings";
 
 import { LanguagePicker } from "@/app/components/language-picker";
 import { getData } from "../data";
