+++ locales/id/messages.json
@@ -5,7 +5,7 @@
   "ctoaas": "CTO as a Service",
   "saas": "Software as a Service",
   "tdd": "Technical Due Dilligence",
-  "erp": "ERP Implementation",
+  "erp": "Implementasi ERP",
   "cloud_migration": "Cloud Migration",
   "volunteering_platform": "Volunteering Platform",
   "our_services": "Layanan Kami",
