+++ locales/id/tdd.json
@@ -3,7 +3,8 @@
     "heading": "Tech Due Diligence",
     "desc": "Identifikasi Risiko. Temukan Peluang. Bangun dengan Keyakinan."
   },
-
+  "description": "Verifikasi dan nilai kemampuan perusahaan Anda dalam menjalankan roadmap serta mengevaluasi kondisi teknis saat ini.",
+  "bestFor": ["Investor", "Direktur teknologi", "perusahaan modal ventura"],
   "what_is_tdd": {
     "desc": "Tech Due Diligence adalah evaluasi terstruktur terhadap lanskap teknologi suatu organisasi. Dirancang untuk mengungkap risiko tersembunyi, menilai kesiapan, dan memberikan kejelasan untuk pengambilan keputusan yang tepat.",
     "highlight": "Baik Anda berinvestasi di startup, bersiap untuk merger, atau mengaudit kapabilitas internal, layanan TDD kami memberikan wawasan yang dapat ditindaklanjuti untuk meminimalkan risiko dalam perjalanan teknologi Anda."
