+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "hyperjump.tech-2024",
-  "version": "0.1.0",
+  "name": "hyperjump.tech",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "generate-locales": "simple-i18n-next -i ./locales -l en",
@@ -15,48 +15,48 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@mdx-js/loader": "3.1.0",
-    "@mdx-js/react": "3.1.0",
-    "@n8n/chat": "^0.27.1",
-    "@next/mdx": "15.3.2",
-    "@next/third-parties": "15.3.2",
-    "@radix-ui/react-accordion": "^1.2.1",
-    "@radix-ui/react-aspect-ratio": "^1.1.2",
-    "@radix-ui/react-avatar": "^1.1.3",
-    "@radix-ui/react-context-menu": "^2.2.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.16",
-    "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-navigation-menu": "^1.2.5",
-    "@radix-ui/react-separator": "^1.1.2",
-    "@radix-ui/react-slot": "^1.1.2",
-    "@radix-ui/react-toast": "^1.2.2",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "framer-motion": "^12.7.4",
-    "lucide-react": "^0.451.0",
-    "next": "15.4.1",
+    "@mdx-js/loader": "3.1.1",
+    "@mdx-js/react": "3.1.1",
+    "@n8n/chat": "0.58.0",
+    "@next/mdx": "15.5.4",
+    "@next/third-parties": "15.5.4",
+    "@radix-ui/react-accordion": "1.2.12",
+    "@radix-ui/react-aspect-ratio": "1.1.7",
+    "@radix-ui/react-avatar": "1.1.10",
+    "@radix-ui/react-context-menu": "2.2.16",
+    "@radix-ui/react-dropdown-menu": "2.1.16",
+    "@radix-ui/react-icons": "1.3.2",
+    "@radix-ui/react-navigation-menu": "1.2.14",
+    "@radix-ui/react-separator": "1.1.7",
+    "@radix-ui/react-slot": "1.2.3",
+    "@radix-ui/react-toast": "1.2.15",
+    "class-variance-authority": "0.7.1",
+    "clsx": "2.1.1",
+    "framer-motion": "12.23.21",
+    "lucide-react": "0.544.0",
+    "next": "15.5.4",
     "next-themes": "0.4.6",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
+    "react": "19.1.1",
+    "react-dom": "19.1.1",
     "schema-dts": "1.1.5",
-    "sonner": "2.0.3",
-    "tailwindcss-animate": "^1.0.7"
+    "sonner": "2.0.7",
+    "tailwindcss-animate": "1.0.7"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.5",
-    "@tailwindcss/typography": "^0.5.16",
+    "@tailwindcss/postcss": "^4.1.13",
+    "@tailwindcss/typography": "^0.5.19",
     "@types/mdx": "^2.0.13",
-    "@types/node": "^22",
-    "@types/react": "19.1.3",
-    "@types/react-dom": "19.1.3",
+    "@types/node": "^24",
+    "@types/react": "19.1.13",
+    "@types/react-dom": "19.1.9",
     "eslint": "^8",
-    "eslint-config-next": "15.3.2",
+    "eslint-config-next": "^15.5.4",
     "postcss": "^8",
-    "prettier": "^3.5.3",
-    "prettier-plugin-tailwindcss": "^0.6.11",
-    "simple-i18n-next": "^0.0.25",
-    "tailwind-merge": "^3.2.0",
-    "tailwindcss": "^4.1.5",
+    "prettier": "^3.6.2",
+    "prettier-plugin-tailwindcss": "^0.6.14",
+    "simple-i18n-next": "^0.0.37",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss": "^4.1.13",
     "typescript": "^5"
   },
   "overrides": {
